refactor: Confirm모달과 AlertModal 내용 바뀐거 수정
