+++ components/src/lib/components/AlertModal/AlertModal.stories.tsx
@@ -17,9 +17,6 @@ const meta: Meta<typeof AlertModal> = {
     content: {
       control: 'text',
     },
-    alertActionsWidth: {
-      control: 'number',
-    },
   },
   args: {
     size: 'medium',
@@ -43,23 +40,6 @@ export const Default: Story = {
     size: 'medium',
     title: 'alert 모달 제목',
     content: 'alert 모달 내용입니다.',
-  },
-  render: (args) => {
-    return (
-      <>
-        <Modal.Trigger>alert모달</Modal.Trigger>
-        <AlertModal {...args} />
-      </>
-    );
-  },
-};
-
-export const AlertWidth: Story = {
-  args: {
-    size: 'medium',
-    title: 'alert 모달 제목',
-    content: 'alert 모달 내용입니다.',
-    alertActionsWidth: 172,
   },
   render: (args) => {
     return (
